src/server/requirements.txt

Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
folium==0.17.0
fonttools==4.37.4
fp==0.2
frozenlist==1.4.1
fsspec==2024.2.0
GeoAlchemy2==0.15.2
geographiclib==2.0
geopandas==0.14.4
geopy==2.4.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gpxpy==1.6.2

stripe==10.9.0
